pyproject.toml
[project]
name = "langstream"
version = "0.1.0"
description = "LangStream is a fastAPI web service for AI-powered conversations using LangGraph agents with multiple LLM providers."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
include = ["src/**/*.py"]
exclude = ["__pycache__", "*.pyc", "*.pyo", "*.pyd", "*.pyw", "*.pyz", "*.pyj", "*.pyz.class", "*.pyz.pyc", "*.pyz.pyo", "*.pyz.pyd", "*.pyz.pyw", "*.pyz.pyj", "*.pyz.pyz.class", "*.pyz.pyz.pyc", "*.pyz.pyz.pyo", "*.pyz.pyz.pyd", "*.pyz.pyz.pyw", "*.pyz.pyz.pyj", "*.pyz.pyz.pyz.class", "*.pyz.pyz.pyz.pyc", "*.pyz.pyz.pyz.pyo", "*.pyz.pyz.pyz.pyd", "*.pyz.pyz.pyz.pyw", "*.pyz.pyz.pyz.pyj"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",  # Ignore blank-line-before-class-docstring (conflicts with D211)
    "D213",  # Ignore multi-line-summary-second-line (conflicts with D212) 
    "COM812", # Ignore trailing-comma (conflicts with formatter)
]

[tool.pyright]
include = ["src.langstream"]
